requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
autopep8==1.4
cachetools==3.1.1
certifi==2018.10.15
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
croniter==0.3.30
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-model-utils==3.2.0
djangorestframework==3.8.2
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==19.9.0
hiredis==1.0.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
msgpack==0.6.1
netifaces==0.10.6
newrelic==4.4.1.104
oauthlib==3.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.4.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.1.1
python-dateutil==2.8.0
python-engineio==2.3.2
pytz==2018.5
redis==3.3.8
requests==2.20.0
requests-oauthlib==1.3.0
rq==1.1.0
rq-scheduler==0.9
rsa==4.0
six==1.11.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24
virtualenv==16.0.0
websocket-client==0.54.0
whitenoise==4.1
wrapt==1.10.11
zope.interface==4.6.0
